helpers/requirements.txt
requests>=2.32
psycopg2-binary>=2.9      # PostgreSQL driver with pre-built wheels
python-slugify>=8.0       # tiny helper used in the download script
